# docs/gpt-actions.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Galaxy of Consequence Controller Actions",
    "version": "1.2.0",
    "description": "Authoritative GOC controller. Gameplay turns, HUD reads, checkpoint saves, checkpoint loads, and OOC reconciliation are separate operations. Administrative operations never roll, narrate, generate initiative, award XP, or advance combat."
  },
  "servers": [
    {
      "url": "https://gqznpnxnbfqvzijxjmtm.supabase.co/functions/v1/goc-gpt-controller"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/gpt/load": {
      "post": {
        "operationId": "loadCheckpoint",
        "summary": "Restore exact checkpoint",
        "description": "Use only when the user explicitly asks to load or restore a checkpoint. Restores the selected snapshot exactly and increments persistence revision once. MUST NOT roll, generate initiative, narrate, advance combat, award XP, or infer extra changes. Call getCampaignState immediately before and pass its revision.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["checkpointId", "expectedRevision"],
                "properties": {
                  "checkpointId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "expectedRevision": {
                    "type": "integer",
                    "minimum": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkpoint restored exactly",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "409": {
            "description": "Revision conflict; reread state before retrying"
          }
        }
      }
    },
    "/api/gpt/reconcile": {
      "post": {
        "operationId": "reconcileCampaignState",
        "summary": "Apply explicit OOC state correction",
        "description": "Administrative state correction only. Use for explicit user-directed HP, XP, credits, location, inventory, combat, or condition-track corrections. Apply ONLY supplied fields. Never invent missing values. MUST NOT roll, generate initiative, narrate, award XP, or infer consequences. Call getCampaignState immediately before and pass its revision.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["expectedRevision", "patch"],
                "properties": {
                  "expectedRevision": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "patch": {
                    "type": "object",
                    "additionalProperties": false,
                    "minProperties": 1,
                    "properties": {
                      "level": { "type": "integer", "minimum": 1 },
                      "experience": { "type": "integer", "minimum": 0 },
                      "health": { "type": "integer" },
                      "credits": { "type": "number" },
                      "location": { "type": "string" },
                      "conditionTrack": { "type": "integer" },
                      "inventory": {
                        "type": "array",
                        "items": {}
                      },
                      "combat": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OOC correction persisted without gameplay advancement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "409": {
            "description": "Revision conflict; reread state before retrying"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}